requirement.txt
baidu-aip==2.2.18.0
certifi==2020.4.5.1
chardet==3.0.4
idna==2.9
lxml==4.5.0
PySimpleGUI==4.18.2
python-docx==0.8.10
requests==2.23.0
urllib3==1.25.9
wincertstore==0.2
